+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Indent `esac`, `od`, `fi` on typing 
 - Prioritize single argument functions on completion
 - Live Templates for functions, `case`/`for`/`while`/`if` expressions
+- Wrong argument count inspection
 
 ### Fixed
 - Documentation comment highlight